Move TimeTickChannel to time service since it isn't needed more generally. Fix other build issues.
