+++ ethereum/events/build.gradle
@@ -2,3 +2,10 @@ dependencies {
   implementation project(':infrastructure:events')
   testFixturesImplementation project(':infrastructure:unsigned')
 }
+
+
+publishing {
+  publications {
+    mavenJava(MavenPublication) { artifactId 'ethereum-events' }
+  }
+}
