+++ ethereum/statetransition/build.gradle
@@ -1,8 +1,9 @@
 dependencies {
+  api project(':ethereum:events')
+
   implementation project(':bls')
   implementation project(':data')
   implementation project(':ethereum:core')
-  implementation project(':ethereum:events')
   implementation project(':ethereum:pow:api')
   implementation project(':ethereum:spec')
   implementation project(':ethereum:weaksubjectivity')
