+++ infrastructure/time/build.gradle
@@ -1,4 +1,3 @@
 dependencies {
-  implementation project(':infrastructure:events')
   testFixturesImplementation project(':infrastructure:unsigned')
 }
