+++ services/timer/src/main/java/tech/pegasys/teku/services/timer/ScheduledTimeEvent.java
@@ -17,7 +17,6 @@ import org.quartz.Job;
 import org.quartz.JobDataMap;
 import org.quartz.JobExecutionContext;
 import org.quartz.JobExecutionException;
-import tech.pegasys.teku.infrastructure.time.TimeTickChannel;
 
 public class ScheduledTimeEvent implements Job {
